scripts/requirements.txt
ansible==3.4.0
ansible-base==2.10.15
cffi==1.15.0
colorful==0.5.4
cryptography==35.0.0
ipaddr==2.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
netaddr==0.8.0
packaging==21.0
pathspec==0.9.0
prettyprinter==0.18.0
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
PyYAML==6.0
six==1.16.0
yamllint==1.26.3
